requirements.txt
Django==4.2.*
django-cors-headers==4.6.0
django-countries-plus==2.2.0
django-environ==0.11.2
django-modeltrans==0.8.0
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
hiredis==3.1.0
html2text==2025.4.15
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-magic==0.4.27; sys_platform=='linux'
python-magic-bin==0.4.14; sys_platform=='win32'
redis==5.2.1
